Normalises everything to require for Node
