+++ src/helpers/executeFetchData/index.js
@@ -1,0 +1,2 @@
+const executeFetchData = require('./executeFetchData')
+module.exports = executeFetchData
